% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{district_group}
\alias{district_group}
\title{Democratic share of district}
\usage{
district_group(plans, group)
}
\arguments{
\item{plans}{a \code{redist_plans} object.}

\item{group}{column of \code{plans} containing the group share of each district.}
}
\value{
a matrix
}
\description{
Returns a matrix of precincts by plans, where each entry is the Democratic
share in the district the precinct belongs to in that plan.
}
