% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{plot_cds}
\alias{plot_cds}
\title{Plot Congressional Districts}
\usage{
plot_cds(map, pl, county, abbr, city = FALSE)
}
\arguments{
\item{map}{redist_map object}

\item{pl}{redist_plans object}

\item{county}{unqouted county name}

\item{abbr}{string, state abbreviation}

\item{city}{boolean. Plot cities? Default is FALSE.}
}
\value{
ggplot
}
\description{
Plot Congressional Districts
}
\examples{
#TODO
}
